SMTP mail client
1. Connects to an SMTP server and identifies the user who will be
sending mail. Both of these values will be hard-coded in your Python
code.  I recommend the server on nike, which is close to your VM in the
network (if you like, you can allow the user to override the sender).
2. Acknowledges to the user that the client is connected to the server
(give the server host name or IP address in this output).
3. Conducts a dialog with the user to get the recipient email address,
message subject, and message text of the message. Because the message
text will include newlines, use the convention of a period on a line by
itself as the input terminator.
4. Sends the message and closes the connection with the server.
5. Helpfully handles the case where the SMTP server does not accept the
recipient address.
 